refactor: Controller를 CarController로 변경 후 패키지 이동
