+++ src/main/kotlin/racingcar/Application.kt
@@ -1,8 +1,10 @@
 package racingcar
 
+import racingcar.controller.CarController
+
 fun main() {
     val view = View()
     val model = Model()
-    val controller = Controller(model, view)
+    val controller = CarController(model, view)
     controller.run()
 }
